requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==26.1.0
bitarray==3.8.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
ckzg==2.1.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.6
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==5.2.0
Flask==3.1.3
frozenlist==1.8.0
hexbytes==1.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.1
packaging==26.0
parsimonious==0.10.0
propcache==0.4.1
py-solc-x==2.0.5
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyunormalize==17.0.0
pywin32==311
regex==2026.4.4
requests==2.33.1
rlp==4.1.0
toolz==1.1.0
types-requests==2.33.0.20260402
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
web3==7.15.0
websockets==15.0.1
Werkzeug==3.1.8
yarl==1.23.0